validate_auth, validate_stripe
