+++ functions/src/stripe/utils.ts
@@ -1,0 +1,18 @@
+import * as functions from 'firebase-functions'
+import Stripe from 'stripe'
+
+export const validate_auth = (context: functions.https.CallableContext) => {
+  if (!context.auth) {
+    throw new functions.https.HttpsError('failed-precondition', 'The function must be called while authenticated.')
+  }
+  return context.auth.uid
+}
+
+export const validate_stripe = () => {
+  const STRIPE_SECRET_KEY = functions.config().stripe.secret_key
+  if (!STRIPE_SECRET_KEY) {
+    throw new functions.https.HttpsError('invalid-argument', 'The functions requires STRIPE_SECRET_KEY.')
+  }
+  return new Stripe(STRIPE_SECRET_KEY, { apiVersion: '2020-03-02' })
+}
+
